requirements.txt
customtkinter==5.2.2
requests==2.32.5
pywin32==311
Pillow==11.1.0
darkdetect==0.8.0
